# 2.5-3X perf improvement of Gloo AllReduce by enabling GPU direct