From my profiling results, we are only getting 1.2GB/s Allreduce throughput from Pytorch on ResNet50 on two nodes with IB. But compared with Microbenchmark roofline with the same data size, we can get around 5.7 GB/s.

Looks like CudaAllreduceHalvingDoublingPipelined by default uses the CudaHostWorkspace, which will not turn on the GPUDirect. This pull request enables this support.

With this change, we are now getting around avg. 3GB/s All reduce throughput for ResNet 50 instead of 1.2GB/s with IB, from my benchmarking/profiling