@soumith 2017-09-22
@pytorchbot add to whitelist