@teng-li 2017-09-22
Used a MACRO to make sure that this is enabled only with that MACRO. Since we currently don't have IBVERB enabled by default, will leave this off until when IBVERB is enabled in the build.