@soumith 2017-09-22
thank you!