@pietern 2017-09-22
I recommend making this conditional at runtime instead of at compile time. You can check for GPUDirect support with context->getDevice()->hasGPUDirect(), and of course checking the GLOO_USE_IBVERBS macro from `gloo/config.h`. Then you have one less choice to make at compile time.